refactor(productMessageService): 메시지 포맷과 서비스 분리를 하나로 통합
