+++ src/main/java/store/convenience/product/controller/ProductController.java
@@ -1,7 +1,6 @@
 package store.convenience.product.controller;
 
 import java.util.List;
-import store.convenience.product.service.ProductMessageFormatter;
 import store.convenience.product.service.ProductService;
 import store.global.util.Reader;
 import store.global.util.StoreConstant;
@@ -9,11 +8,9 @@ import store.global.util.StoreConstant;
 public class ProductController {
 
     private final ProductService productService;
-    private final ProductMessageFormatter messageFormatter;
 
-    public ProductController(ProductService productService, ProductMessageFormatter messageFormatter) {
+    public ProductController(ProductService productService) {
         this.productService = productService;
-        this.messageFormatter = messageFormatter;
     }
 
     public void start() {
